feat(question-type-ranking): #FOR-451 update cursor directive
