+++ formulaire/src/main/resources/public/ts/directives/question/question-type/index.ts
@@ -8,5 +8,5 @@ export * from './question-type-time';
 export * from './question-type-file';
 export * from './question-type-singleanswer-radio';
 export * from './question-type-matrix';
-export * from './question-type-cursor';
+export * from './question-type-cursor/question-type-cursor.directive';
 export * from './question-type-ranking/question-type-ranking.directive';
