+++ formulaire/src/main/resources/public/ts/directives/question/question-type/question-type-cursor/question-type-cursor.html
@@ -1,0 +1,54 @@
+<div class="formulaire-cursor-form">
+  <!-- Line about MIN value and label -->
+  <div class="cursor-line">
+    <div class="cursor-value-min twelve-mobile">
+      <div class="six four-mobile"><label><i18n>formulaire.question.value.minimum</i18n><em>*</em> : </label></div>
+      <div ng-if="!vm.hasFormResponses">
+        <input type="number" ng-model="vm.question.cursor_min_val"
+               max="[[vm.question.cursor_max_val - vm.question.cursor_step]]" pattern="\d*"
+               i18n-placeholder="formulaire.question.cursor.default.min.val">
+      </div>
+      <div ng-if="vm.hasFormResponses">
+        <input type="number" disabled ng-model="vm.question.cursor_min_val">
+      </div>
+    </div>
+    <div class="cursor-label-min-val">
+      <label><i18n>formulaire.question.value.label</i18n></label>
+      <input type="text" ng-model="vm.question.cursor_min_label"
+             i18n-placeholder="formulaire.question.label">
+    </div>
+  </div>
+  <!-- Line about MAX value and label -->
+  <div class="cursor-line">
+    <div class="cursor-value-max twelve-mobile">
+      <div class="six four-mobile"><label><i18n>formulaire.question.value.maximum</i18n><em>*</em> : </label></div>
+      <div ng-if="!vm.hasFormResponses">
+        <input type="number" ng-model="vm.question.cursor_max_val"
+               min="[[vm.question.cursor_min_val + vm.question.cursor_step]]" pattern="\d*"
+               i18n-placeholder="formulaire.question.cursor.default.max.val">
+      </div>
+      <div ng-if="vm.hasFormResponses">
+        <input type="number" disabled ng-model="vm.question.cursor_min_val">
+      </div>
+    </div>
+    <div class="cursor-label-max-value">
+      <label><i18n>formulaire.question.value.label</i18n></label>
+      <input type="text" ng-model="vm.question.cursor_max_label"
+             i18n-placeholder="formulaire.question.label">
+    </div>
+  </div>
+  <!-- Line about STEP value -->
+  <div class="cursor-line">
+    <div class="cursor-value-step">
+      <div class="six four-mobile"><label><i18n>formulaire.question.value.step</i18n><em>*</em> : </label></div>
+      <div ng-if="!vm.hasFormResponses">
+        <input type="number" ng-model="vm.question.cursor_step"
+               ng-change="vm.onChangeStep(vm.question.cursor_step)" pattern="\d*"
+               i18n-placeholder="formulaire.question.cursor.default.step">
+      </div>
+      <div ng-if="vm.hasFormResponses">
+        <input type="number" disabled ng-model="vm.question.cursor_step">
+      </div>
+    </div>
+  </div>
+</div>
